+++ src/router/main.routes.ts
@@ -13,6 +13,7 @@ import AccessDenied from "@/auth/AccessDenied.vue";
 import NotFound from "@/page/NotFound.vue";
 import corpusRoutes from "@/router/corpus.routes";
 import metadataRoutes from "@/router/metadata.routes";
+import lexiconRoutes from "@/router/lexicon.routes";
 
 const routes: RouteRecordRaw[] = [
   {
@@ -77,6 +78,7 @@ const routes: RouteRecordRaw[] = [
     },
   },
   ...corpusRoutes,
+  ...lexiconRoutes,
   ...metadataRoutes,
   {
     path: "/:pathMatch(.*)*",